+++ CHANGELOG.md
@@ -4,10 +4,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
-- Use ContentMD5Value parameter instead of Content-MD5 header.
 - Parse MWS headers.
 - Delegate to Hash#dig in Peddler::XMLParser.
 - Add InvalidRequest and MalformedInput to known error codes.
+
+### Changed
+- Use ContentMD5Value parameter instead of Content-MD5 header.
 
 ### Fixed
 - Post multibyte characters correctly.